clean up patients controller
