--- src/IbdTracker/Features/Patients/GetAll.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using IbdTracker.Core;
-using IbdTracker.Core.CommonDtos;
-using MediatR;
-
-namespace IbdTracker.Features.Patients
-{
-    public class GetAll
-    {
-        public class Query : IRequest<IList<PatientDto>>
-        {
-        }
-        
-        public class Handler : IRequestHandler<Query, IList<PatientDto>>
-        {
-            private readonly IbdSymptomTrackerContext _context;
-
-            public Handler(IbdSymptomTrackerContext context)
-            {
-                _context = context;
-            }
-
-            public async Task<IList<PatientDto>> Handle(Query request, CancellationToken cancellationToken)
-            {
-                throw new NotImplementedException();
-            }
-        }
-    }
-}
